cancro: use open source thermanager
The current thermal solution spams the shit out of logcat
and doesn't really work anyway. Sony's thermanager is open
and proven to be effective.

This config is based on Xperia Z3 config, but with limits
approximately +10'C to compensate for Xiaomi's vastly inferior
hardware heat dispersal expecially on the mi4.

Change-Id: I41ce29e1e92d930e0f79590a39fff08246b2d90e
Signed-off-by: jrizzoli <joey@cyanogenmoditalia.it>