cancro: healtd: the worst thing i could do
Since the day I switched to healtd many humans
(and softwares) noticed that device wasn't charging
as fast as it was excepted (in fact hevc is enabled).
After years of studies sciencists realized that
it was always charging less than 220mah.
The problem magically disappeared if the device booted while it
was connected to the charger in "off-mode" or healtd was nuked.
This issue made users so sad that many of them could not sleep at
night and they were not able to be productive at work anymore,
so they got fired...
Hopefully I've found out a way to use xiaomi's
chargeonlymode with our new MmMmMmarshmallowish android release.

TL;DR: switch back to xiaomi's chargeonlymode

Change-Id: I165f9e090db30de101abe758c49041cd65c29299
Signed-off-by: jrizzoli <joey@cyanogenmoditalia.it>